--- budget-binder-multiplatform-app/src/commonMain/kotlin/de/hsfl/budgetBinder/domain/usecase/DataFlowUseCases.kt
@@ -1,11 +1,0 @@
-package de.hsfl.budgetBinder.domain.usecase
-
-import de.hsfl.budgetBinder.domain.usecase.storage.StoreDarkModeUseCase
-import de.hsfl.budgetBinder.domain.usecase.storage.StoreServerUrlUseCase
-import de.hsfl.budgetBinder.domain.usecase.storage.StoreUserStateUseCase
-
-data class DataFlowUseCases(
-    val storeUserStateUseCase: StoreUserStateUseCase,
-    val storeServerUrlUseCase: StoreServerUrlUseCase,
-    val storeDarkModeUseCase: StoreDarkModeUseCase
-)
