--- budget-binder-multiplatform-app/src/commonMain/kotlin/de/hsfl/budgetBinder/domain/usecase/storage/StoreDarkModeUseCase.kt
@@ -1,10 +1,0 @@
-package de.hsfl.budgetBinder.domain.usecase.storage
-
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.flow
-
-class StoreDarkModeUseCase {
-    operator fun invoke(isDarkMode: Boolean): Flow<Boolean> = flow {
-        emit(isDarkMode)
-    }
-}
