+++ budget-binder-multiplatform-app/src/commonMain/kotlin/de/hsfl/budgetBinder/presentation/viewmodel/settings/SettingsEditUserViewModel.kt
@@ -5,7 +5,7 @@ import de.hsfl.budgetBinder.domain.usecase.SettingsUseCases
 import de.hsfl.budgetBinder.presentation.Screen
 import de.hsfl.budgetBinder.presentation.event.UiEvent
 import de.hsfl.budgetBinder.presentation.flow.DarkModeFlow
-import de.hsfl.budgetBinder.presentation.flow.DataFlow
+import de.hsfl.budgetBinder.presentation.flow.UserFlow
 import de.hsfl.budgetBinder.presentation.flow.RouterFlow
 import de.hsfl.budgetBinder.presentation.flow.UiEventSharedFlow
 import kotlinx.coroutines.CoroutineScope
@@ -15,7 +15,7 @@ import kotlinx.coroutines.launch
 
 class SettingsEditUserViewModel(
     private val settingsUseCases: SettingsUseCases,
-    private val dataFlow: DataFlow,
+    private val userFlow: UserFlow,
     private val routerFlow: RouterFlow,
     private val scope: CoroutineScope,
     darkModeFlow: DarkModeFlow
@@ -36,14 +36,14 @@ class SettingsEditUserViewModel(
     private val _confirmedPasswordText = MutableStateFlow(EditUserState())
     val confirmedPassword: StateFlow<EditUserState> = _confirmedPasswordText
 
-    val emailText: StateFlow<String> = MutableStateFlow(dataFlow.userState.value.email)
+    val emailText: StateFlow<String> = MutableStateFlow(userFlow.userState.value.email)
 
     private val _eventFlow = UiEventSharedFlow.mutableEventFlow
 
     init {
         _firstNameText.value =
-            firstNameText.value.copy(firstName = dataFlow.userState.value.firstName, firstNameIsValid = true)
-        _lastNameText.value = lastNameText.value.copy(lastName = dataFlow.userState.value.name, lastNameIsValid = true)
+            firstNameText.value.copy(firstName = userFlow.userState.value.firstName, firstNameIsValid = true)
+        _lastNameText.value = lastNameText.value.copy(lastName = userFlow.userState.value.name, lastNameIsValid = true)
         _passwordText.value = passwordText.value.copy(password = passwordPlaceholder, passwordIsValid = true)
         _confirmedPasswordText.value =
             confirmedPassword.value.copy(confirmedPassword = passwordPlaceholder, confirmedPasswordIsValid = true)
@@ -124,7 +124,7 @@ class SettingsEditUserViewModel(
                     logOutOnAllDevices("Your password was updated. Please sign in again")
                 } else {
                     _eventFlow.emit(UiEvent.ShowSuccess("User update successfully"))
-                    dataFlow.storeUserState(it)
+                    userFlow.storeUserState(it)
                     routerFlow.navigateTo(Screen.Settings.Menu)
                 }
             })
@@ -145,8 +145,8 @@ class SettingsEditUserViewModel(
 
     private fun resetFlows() {
         _firstNameText.value =
-            firstNameText.value.copy(firstName = dataFlow.userState.value.firstName, firstNameIsValid = true)
-        _lastNameText.value = lastNameText.value.copy(lastName = dataFlow.userState.value.name, lastNameIsValid = true)
+            firstNameText.value.copy(firstName = userFlow.userState.value.firstName, firstNameIsValid = true)
+        _lastNameText.value = lastNameText.value.copy(lastName = userFlow.userState.value.name, lastNameIsValid = true)
         _passwordText.value = passwordText.value.copy(password = passwordPlaceholder, passwordIsValid = true)
     }
 }
